# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Financial Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface BudgetCategory {
  id: number;
  name: string;
  year: number;
  initialBudget: number;
  finalBudget: number;
  spent: number;
  remaining: number;
  utilizationPct: number;
  createdAt: string;
}

export interface CreateBudgetCategory {
  name: string;
  year: number;
  initialBudget: number;
  finalBudget: number;
  spent: number;
  remaining: number;
  utilizationPct: number;
}

export interface BudgetLineItem {
  id: number;
  categoryId: number;
  name: string;
  amount: number;
  percentage: number;
  year: number;
  createdAt: string;
}

export interface CreateBudgetLineItem {
  categoryId: number;
  name: string;
  amount: number;
  percentage: number;
  year: number;
}

export interface YearlyComparison {
  id: number;
  year: number;
  totalBudget: number;
  totalSpent: number;
  utilizationPct: number;
}

export interface CreateYearlyComparison {
  year: number;
  totalBudget: number;
  totalSpent: number;
  utilizationPct: number;
}

export interface OwnRevenue {
  id: number;
  name: string;
  amount: number;
  percentage: number;
  year: number;
  createdAt: string;
}

export interface CreateOwnRevenue {
  name: string;
  amount: number;
  percentage: number;
  year: number;
}

export type BalanceSheetItemCategory = typeof BalanceSheetItemCategory[keyof typeof BalanceSheetItemCategory];


export const BalanceSheetItemCategory = {
  asset: 'asset',
  liability: 'liability',
  capital: 'capital',
} as const;

export interface BalanceSheetItem {
  id: number;
  category: BalanceSheetItemCategory;
  name: string;
  amount: number;
  year: number;
  sortOrder: number;
}

export interface BudgetUtilizationSummary {
  year: number;
  totalInitialBudget: number;
  totalFinalBudget: number;
  totalSpent: number;
  totalRemaining: number;
  overallUtilizationPct: number;
  categories: BudgetCategory[];
}

export type CustomDatasetChartType = typeof CustomDatasetChartType[keyof typeof CustomDatasetChartType];


export const CustomDatasetChartType = {
  bar: 'bar',
  line: 'line',
  pie: 'pie',
  area: 'area',
} as const;

export interface CustomDataset {
  id: number;
  name: string;
  description?: string;
  chartType: CustomDatasetChartType;
  xAxisLabel?: string;
  yAxisLabel?: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateCustomDatasetChartType = typeof CreateCustomDatasetChartType[keyof typeof CreateCustomDatasetChartType];


export const CreateCustomDatasetChartType = {
  bar: 'bar',
  line: 'line',
  pie: 'pie',
  area: 'area',
} as const;

export interface CreateCustomDataset {
  name: string;
  description?: string;
  chartType: CreateCustomDatasetChartType;
  xAxisLabel?: string;
  yAxisLabel?: string;
}

export type CustomDatasetWithRowsChartType = typeof CustomDatasetWithRowsChartType[keyof typeof CustomDatasetWithRowsChartType];


export const CustomDatasetWithRowsChartType = {
  bar: 'bar',
  line: 'line',
  pie: 'pie',
  area: 'area',
} as const;

export interface CustomDataRow {
  id: number;
  datasetId: number;
  label: string;
  value: number;
  color?: string;
  sortOrder: number;
  createdAt: string;
}

export interface CustomDatasetWithRows {
  id: number;
  name: string;
  description?: string;
  chartType: CustomDatasetWithRowsChartType;
  xAxisLabel?: string;
  yAxisLabel?: string;
  createdAt: string;
  updatedAt: string;
  rows: CustomDataRow[];
}

export interface CreateCustomDataRow {
  label: string;
  value: number;
  color?: string;
  sortOrder: number;
}

export type ListBudgetCategoriesParams = {
year?: number;
};

export type ListBudgetLineItemsParams = {
categoryId?: number;
year?: number;
};

export type ListOwnRevenuesParams = {
year?: number;
};

export type ListBalanceSheetItemsParams = {
year?: number;
};

export type GetBudgetUtilizationSummaryParams = {
year?: number;
};