# lib/api-zod/src/generated/types/budgetCategory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Financial Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface BudgetCategory {
  id: number;
  name: string;
  year: number;
  initialBudget: number;
  finalBudget: number;
  spent: number;
  remaining: number;
  utilizationPct: number;
  createdAt: Date;
}