# lib/api-zod/src/generated/types/createCustomDataset.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Financial Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCustomDatasetChartType } from './createCustomDatasetChartType';

export interface CreateCustomDataset {
  name: string;
  description?: string;
  chartType: CreateCustomDatasetChartType;
  xAxisLabel?: string;
  yAxisLabel?: string;
}