# lib/api-zod/src/generated/types/customDataRow.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Financial Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomDataRow {
  id: number;
  datasetId: number;
  label: string;
  value: number;
  color?: string;
  sortOrder: number;
  createdAt: Date;
}