# lib/api-zod/src/generated/types/customDatasetWithRows.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Financial Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomDataRow } from './customDataRow';
import type { CustomDatasetWithRowsChartType } from './customDatasetWithRowsChartType';

export interface CustomDatasetWithRows {
  id: number;
  name: string;
  description?: string;
  chartType: CustomDatasetWithRowsChartType;
  xAxisLabel?: string;
  yAxisLabel?: string;
  createdAt: Date;
  updatedAt: Date;
  rows: CustomDataRow[];
}